+++ test/test_gmac.c
@@ -63,7 +63,7 @@ static int test_gmac_gen_mac(OSSL_LIB_CTX* libCtx, const char *c,
         err = (EVP_MAC_final(mctx, out, &outSz, *outLen)) != 1;
     }
     if (err == 0) {
-        *outLen = outSz;
+        *outLen = (int)outSz;
         PRINT_BUFFER("GMAC", out, *outLen);
     }
